# [2/8][AccountManager]: benchmark `ListAccounts` API against deprecated `Accounts` wrapper API