--- a/wallet/account_manager_benchmark_test.go
+++ b/wallet/account_manager_benchmark_test.go
@@ -60,3 +60,55 @@ func BenchmarkListAccountsByScopeAPI(b *testing.B) {
 		})
 	}
 }
+
+// BenchmarkListAccountsAPI benchmarks ListAccounts API and a deprecated variant
+// of it using same key scopes and identical test data across multiple dataset
+// sizes. Test names start with dataset size to group API comparisons for
+// benchstat analysis.
+func BenchmarkListAccountsAPI(b *testing.B) {
+	benchmarkSizes, namingInfo := generateBenchmarkSizes(
+		benchmarkConfig{
+			accountGrowth: linearGrowth,
+			utxoGrowth:    exponentialGrowth,
+			maxIterations: 14,
+			startIndex:    0,
+		},
+	)
+	scopes := waddrmgr.DefaultKeyScopes
+
+	for _, size := range benchmarkSizes {
+		b.Run(size.name(namingInfo)+"/0-Before", func(b *testing.B) {
+			w := setupBenchmarkWallet(
+				b, benchmarkWalletConfig{
+					scopes:      scopes,
+					numAccounts: size.numAccounts,
+					numUTXOs:    size.numUTXOs,
+				},
+			)
+
+			b.ReportAllocs()
+			b.ResetTimer()
+			for b.Loop() {
+				_, err := listAccountsDeprecated(w)
+				require.NoError(b, err)
+			}
+		})
+
+		b.Run(size.name(namingInfo)+"/1-After", func(b *testing.B) {
+			w := setupBenchmarkWallet(
+				b, benchmarkWalletConfig{
+					scopes:      scopes,
+					numAccounts: size.numAccounts,
+					numUTXOs:    size.numUTXOs,
+				},
+			)
+
+			b.ReportAllocs()
+			b.ResetTimer()
+			for b.Loop() {
+				_, err := w.ListAccounts(b.Context())
+				require.NoError(b, err)
+			}
+		})
+	}
+}